Move to exposure history to view layer
Why:
Currently we are setting the root data model for exposures as an
ExposureHistory. We would like to set the root data model to be a less
view specific data shape, ExposureInfo, so that we can better separate
the data model and the presentational layers.

This commit:
Converts the ExposureHistoryContext to an ExposureContext and moves the
data converters from exposureHistory.ts to the gaen folder for
the conversion of raw exposure to the application target of an
ExposureInfo and the converter from the ExposureInfo target to a user
facing ExposureHistory to the ExposureHistory folder.

We were also able to move the native module call getLastDectionDate
into the tracingStrategy so that the IO is located at shell of the
system and can be better injected as a dependency so that it is easier
to test the system.
