+++ src/ExposureHistory/Calendar.tsx
@@ -3,7 +3,8 @@ import { Text, TouchableOpacity, View, StyleSheet } from "react-native"
 import { useTranslation } from "react-i18next"
 import dayjs from "dayjs"
 
-import { ExposureHistory, ExposureDatum } from "../exposureHistory"
+import { ExposureHistory } from "./exposureHistory"
+import { ExposureDatum } from "../exposure"
 import { RTLEnabledText } from "../components/RTLEnabledText"
 import ExposureDatumIndicator from "./ExposureDatumIndicator"
 import LegendModal from "./LegendModal"
