+++ src/ExposureHistory/DateInfoHeader.tsx
@@ -1,4 +1,4 @@
-import React, { FunctionComponent, useContext } from "react"
+import React, { FunctionComponent } from "react"
 import { StyleSheet } from "react-native"
 import { useTranslation } from "react-i18next"
 import dayjs from "dayjs"
@@ -6,26 +6,31 @@ import relativeTime from "dayjs/plugin/relativeTime"
 dayjs.extend(relativeTime)
 
 import { RTLEnabledText } from "../components/RTLEnabledText"
-import ExposureHistoryContext from "../ExposureHistoryContext"
+import { DateTimeUtils } from "../utils"
 
 import { Typography } from "../styles"
 
-const DateInfoHeader: FunctionComponent = () => {
-  const { t } = useTranslation()
-  const { lastExposureDetectionDate } = useContext(ExposureHistoryContext)
+type Posix = number
 
-  if (lastExposureDetectionDate === null) {
-    return null
+interface DateInfoHeaderProps {
+  lastDetectionDate: Posix | null
+}
+
+const DateInfoHeader: FunctionComponent<DateInfoHeaderProps> = ({
+  lastDetectionDate,
+}) => {
+  const { t } = useTranslation()
+
+  const determineUpdatedAtText = (posix: Posix): string => {
+    const updated = t("exposure_history.updated")
+    const timeAgoInWords = DateTimeUtils.timeAgoInWords(posix)
+    return ` • ${updated} ${timeAgoInWords}`
   }
 
   const lastDaysText = t("exposure_history.last_days")
-  const updated = t("exposure_history.updated")
-  let updatedAtText = ""
-
-  if (lastExposureDetectionDate !== null) {
-    const humanizedTimePassed = lastExposureDetectionDate.fromNow()
-    updatedAtText = ` • ${updated} ${humanizedTimePassed}`
-  }
+  const updatedAtText = lastDetectionDate
+    ? determineUpdatedAtText(lastDetectionDate)
+    : ""
 
   return (
     <RTLEnabledText style={styles.subHeaderText}>
