+++ src/factories/rawExposure.tsx
@@ -1,7 +1,7 @@
 import { Factory } from "fishery"
 
 import { DateTimeUtils } from "../utils"
-import { RawExposure } from "../gaen/exposureNotifications"
+import { RawExposure } from "../gaen/dataConverters"
 
 export default Factory.define<RawExposure>(() => {
   const defaultDate = DateTimeUtils.beginningOfDay(DateTimeUtils.daysAgo(2))
