+++ src/gaen/dataConverter.spec.ts
@@ -1,77 +1,8 @@
 import dayjs from "dayjs"
 
 import { DateTimeUtils } from "../utils"
-import { Possible, ExposureDatum, ExposureInfo } from "../exposureHistory"
-import {
-  toExposureInfo,
-  RawExposure,
-  toExposureHistory,
-} from "./exposureNotifications"
-
-describe("toExposureHistory", () => {
-  describe("when given a 21 day calendar", () => {
-    describe("when given an ExposureInfo with no exposures", () => {
-      it("it returns a 21 day history of NoKnown Exposures", () => {
-        const today = dayjs("2020-06-26").valueOf()
-        const exposureInfo: ExposureInfo = {}
-
-        const result = toExposureHistory(exposureInfo, {
-          startDate: today,
-          totalDays: 21,
-        })
-
-        expect(result.length).toBe(21)
-        expect(
-          result.every((datum: ExposureDatum) => datum.kind === "NoKnown"),
-        ).toBe(true)
-      })
-    })
-
-    describe("when given an ExposureInfo with a possible exposure today", () => {
-      it("returns an exposure history with the exposures on the correct days", () => {
-        const tuesday = dayjs("2020-06-23").valueOf()
-        const friday = dayjs("2020-06-26").valueOf()
-
-        const today = friday
-
-        const tuesdayExposure: ExposureDatum = {
-          kind: "Possible",
-          date: tuesday,
-          duration: 600000,
-          totalRiskScore: 3,
-          transmissionRiskLevel: 3,
-        }
-        const fridayExposure: ExposureDatum = {
-          kind: "Possible",
-          date: friday,
-          duration: 300000,
-          totalRiskScore: 4,
-          transmissionRiskLevel: 7,
-        }
-
-        const exposureInfo: ExposureInfo = {
-          [tuesday]: tuesdayExposure,
-          [friday]: fridayExposure,
-        }
-
-        const result = toExposureHistory(exposureInfo, {
-          startDate: today,
-          totalDays: 21,
-        })
-
-        const tuesdayResult = result[result.length - 5]
-        const wednesdayResult = result[result.length - 4]
-        const fridayResult = result[result.length - 2]
-        const saturdayResult = result[result.length - 1]
-
-        expect(tuesdayResult).toEqual(tuesdayExposure)
-        expect(wednesdayResult.kind).toBe("NoKnown")
-        expect(fridayResult).toEqual(fridayExposure)
-        expect(saturdayResult.kind).toBe("NoKnown")
-      })
-    })
-  })
-})
+import { Possible } from "../exposure"
+import { toExposureInfo, RawExposure } from "./dataConverters"
 
 describe("toExposureInfo", () => {
   describe("when there are no exposure notifications", () => {
